split mean field in submodules
