+++ alpsqutip/operators/states/meanfield/projections.py
@@ -2,15 +2,14 @@
 Module that implements a meanfield approximation of a Gibbsian state
 """
 
-import logging
+
 from functools import reduce
 from itertools import combinations
 from typing import Optional, Tuple, Union
 
-import numpy as np
+
 import qutip
 from qutip import Qobj
-from scipy.optimize import minimize_scalar
 
 from alpsqutip.operators import (
     LocalOperator,
@@ -25,7 +24,6 @@ from alpsqutip.operators.states.basic import (
     DensityOperatorMixin,
     ProductDensityOperator,
 )
-from alpsqutip.operators.states.gibbs import GibbsProductDensityOperator
 from alpsqutip.qutip_tools.tools import schmidt_dec_firsts_last_qutip_operator
 from alpsqutip.settings import ALPSQUTIP_TOLERANCE
 
@@ -147,7 +145,7 @@ def one_body_from_qutip_operator(
     )
 
 
-def project_operator_to_m_body(full_operator: Operator, m_max=2, sigma_0=None):
+def project_operator_to_m_body(full_operator: Operator, m_max=2, sigma_0=None)->Operator:
     """
     Project a Operator onto a m_max - body operators sub-algebra
     relative to the local states `local_sigmas`.
@@ -223,7 +221,7 @@ def project_operator_to_m_body(full_operator: Operator, m_max=2, sigma_0=None):
     )
 
 
-def project_qutip_operator_to_m_body(full_operator: Operator, m_max=2, sigma_0=None):
+def project_qutip_operator_to_m_body(full_operator: Operator, m_max=2, sigma_0=None)->Operator:
     """
     Specialized version for QutipOperators.
     """
@@ -289,68 +287,11 @@ def project_qutip_operator_to_m_body(full_operator: Operator, m_max=2, sigma_0=N
     return ScalarOperator(0, full_operator.system)
 
 
-def self_consistent_quadratic_mfa(ham: Operator):
-    """
-    Find the Mean field approximation for the exponential
-    of a quadratic form.
-
-    Starts by decomposing ham as a quadratic form
-
-    """
-    from alpsqutip.operators.quadratic import build_quadratic_form_from_operator
-
-    system = ham.system
-    ham_qf = build_quadratic_form_from_operator(ham, isherm=True, simplify=True)
-    # TODO: use random choice
-    basis = sorted(
-        [(w, b) for w, b in zip(ham_qf.weights, ham_qf.basis) if w < 0],
-        key=lambda x: x[0],
-    )
-    w_0, b_0 = basis[0]
-    offset = ham_qf.offset or ScalarOperator(0, system)
-
-    def try_state(beta):
-        k_op = beta * w_0 * b_0 + offset
-        return GibbsProductDensityOperator(k_op, system=b_0.system), k_op
-
-    def hartree_free_energy(state, k_op):
-        free_energy = np.real(sum(state.free_energies.values()))
-        h_av, k_av = np.real(state.expect([state.expect(ham), state.expect(k_op)]))
-        return np.real(h_av - k_av + free_energy)
-
-    # Start by optimizing with the best candidate:
-    def try_function(beta):
-        return hartree_free_energy(*try_state(beta)) + 0.001 * beta**2
-
-    res = minimize_scalar(try_function, (-0.2, 0.25), method="golden")
-    # Now, run a self consistent loop
-    h_fe = res.fun
-    phis = [0 for b in basis]
-    phis[0] = res.x
-    state_mf, kappa = try_state(res.x)
-
-    for _ in range(10):
-        exp_vals = state_mf.expect([w_b[1] for w_b in basis]).real
-        new_phis = [2 * o_av for w_b, o_av in zip(basis, exp_vals)]
-        phis = new_phis
-        # phis = [.75*a + .25*b for a,b in zip(phis, new_phis)]
-        kappa = offset + sum(
-            w_b[1] * (w_b[0] * coeff) for coeff, w_b in zip(phis, basis)
-        )
-        kappa = kappa.simplify()
-        state_mf = GibbsProductDensityOperator(kappa, system=system)
-        new_h_fe = hartree_free_energy(state_mf, kappa)
-        if new_h_fe > h_fe:
-            break
-        h_fe = new_h_fe
-    return state_mf
-
-
 def project_product_operator_as_n_body_operator(
     operator: ProductOperator,
     nmax: Optional[int] = 1,
     sigma: Optional[ProductDensityOperator] = None,
-):
+)->Operator:
     """
     Project a product operator to the manifold of n-body operators
     """
@@ -395,7 +336,7 @@ def project_product_operator_as_n_body_operator(
 
 def project_quadraticform_operator_as_n_body_operator(
     operator, nmax: Optional[int] = 1, sigma: Optional[ProductDensityOperator] = None
-):
+)->Operator:
     """
     Project a product operator to the manifold of n-body operators
     """
@@ -415,7 +356,7 @@ def project_quadraticform_operator_as_n_body_operator(
 
 def project_qutip_operator_as_n_body_operator(
     operator, nmax: Optional[int] = 1, sigma: Optional[ProductDensityOperator] = None
-):
+)->Operator:
     """
     Project a qutip operator to the manifold of n-body operators
     """
@@ -474,7 +415,7 @@ def project_qutip_operator_as_n_body_operator(
     return SumOperator(tuple(terms), system)
 
 
-def project_to_n_body_operator(operator, nmax=1, sigma=None):
+def project_to_n_body_operator(operator, nmax=1, sigma=None)->Operator:
     """
     Approximate `operator` by a sum of (up to) nmax-body
     terms, relative to the state sigma.
@@ -573,80 +514,3 @@ def project_to_n_body_operator(operator, nmax=1, sigma=None):
     return SumOperator(tuple(terms), system)
 
 
-def project_meanfield(
-    k_op, sigma0=None, max_it=100, proj_func=project_operator_to_m_body
-):
-    """
-    Look for a one-body operator kmf s.t
-    Tr (k_op-kmf)exp(-kmf)=0
-
-    following a self-consistent, iterative process
-    assuming that exp(-kmf)~sigma0
-
-    If sigma0 is not provided, sigma0 is taken as the
-    maximally mixed state.
-
-    """
-    sigma0 = self_consistent_project_meanfield(
-        k_op, sigma0, max_it, proj_func=proj_func
-    )[1]
-    result = proj_func(k_op, 1, sigma0).simplify()
-    # result = project_to_n_body_operator(k_op, 1, sigma0).simplify()
-    return result
-
-
-def self_consistent_project_meanfield(
-    k_op, sigma=None, max_it=100, proj_func=project_operator_to_m_body
-) -> Tuple[Operator, Operator]:
-    """
-    Iteratively computes the one-body component from a QuTip operator and state
-    using a self-consistent Mean-Field Projection (MF).
-
-    Parameters:
-        k_op: The initial operator, a QuTip.Qobj, to be decomposed into
-        one-body components.
-        sigma: The referential state to be used as the initial guess
-               in the calculations.
-        k_0: if given, the logarithm of sigma.
-        max_it: Maximum number of iterations.
-
-    Returns:
-        A tuple (K_one_body, sigma_one_body):
-        - K_one_body: The one-body component of the operator K, an
-        AlpsQuTip.one_body_operator object.
-        - sigma_one_body: The one-body state normalized through the
-        MFT process.
-    """
-    if sigma is None:
-        sigma = GibbsProductDensityOperator(k={}, system=k_op.system)
-        neg_log_sigma = -sigma.logm()
-    else:
-        neg_log_sigma = -sigma.logm()
-        if not isinstance(sigma, GibbsProductDensityOperator):
-            sigma = GibbsProductDensityOperator(neg_log_sigma)
-
-    rel_s = 10000
-    opt_sigma = sigma
-
-    for it in range(max_it):
-        # k_one_body = project_operator_to_m_body(k_op, 1, sigma)
-        k_one_body = project_to_n_body_operator(k_op, 1, sigma).simplify()
-        new_sigma = GibbsProductDensityOperator(k_one_body)
-
-        log_k_one_body = new_sigma.logm()
-        rel_s_new = np.real(sigma.expect(k_op + log_k_one_body))
-        rel_entropy_txt = f"     S(curr||target)={rel_s_new}"
-        logging.debug(rel_entropy_txt)
-        # print(rel_entropy_txt)
-        if it > 20 and rel_s_new > 2 * rel_s:
-            break
-
-        if rel_s_new < rel_s:
-            rel_s = rel_s_new
-            opt_sigma = new_sigma
-            sigma = new_sigma
-        else:
-            sigma = new_sigma
-
-    k_one_body = project_to_n_body_operator(k_op, 1, opt_sigma).simplify()
-    return k_one_body, opt_sigma
